pedidos

+++ src/components/views/producto/CardProducto.jsx
@@ -1,23 +1,35 @@
 import { Col, Card } from "react-bootstrap";
-
 import { Link,useNavigate} from "react-router-dom";
 import React from "react";
 import "../paginaprincipal.css";
+import Swal from "sweetalert2";
 
 const CardProducto = ({ user,producto }) => {
   const navigate = useNavigate();
   console.log(user)
   const handleComprarClick = () => {
-    if(user===null){
-      navigate("/login")
-    }else{  
+    if (user === null) {
+      navigate("/login");
+    } else {
       const productoJSON = JSON.stringify(producto);
       const productosSeleccionados = JSON.parse(localStorage.getItem("productosSeleccionados")) || [];
-      productosSeleccionados.push(producto);
-      localStorage.setItem("productosSeleccionados", JSON.stringify(productosSeleccionados));
-      navigate(`/pedidos/${producto._id}`)
+  
+      // Verificar si el producto ya está en el carrito
+      const productoExistente = productosSeleccionados.find(item => item._id === producto._id);
+  
+      if (productoExistente) {
+        // Mostrar mensaje de error usando SweetAlert2
+        Swal.fire({
+          icon: 'error',
+          title: 'Producto ya en el carrito',
+          text: 'Este producto ya ha sido agregado al carrito.',
+        });
+      } else {
+        productosSeleccionados.push(producto);
+        localStorage.setItem("productosSeleccionados", JSON.stringify(productosSeleccionados));
+        navigate(`/pedidos/${producto._id}`);
+      }
     }
-    
   };
